Dimming: new architecture allowing device hardware specific optimisation (and a generic base class implementation for simple devices). Tested with hue, DALI and simulated devices.

+++ src/deviceclasses/simpleio/consoledevice.hpp
@@ -40,7 +40,7 @@ namespace p44 {
     bool hasOutput;
     bool hasColor;
     ConsoleKeyPtr consoleKey;
-    int32_t outputValue;
+    double outputValue;
 
   public:
     ConsoleDevice(StaticDeviceContainer *aClassContainerP, const string &aDeviceConfig);
